Type fix part 2
Map ID needs to be int. Overall ratings need to be float.

+++ api/persistence/implementations/washroom_impl.py
@@ -36,7 +36,7 @@ class WashroomsPersistence(IWashroomsPersistence):
         floor: int,
         gender: str,
         amenities_id: int,  # Foreign Key
-        overall_rating: int,
+        overall_rating: float,
         average_ratings_id: int  # Foreign Key
     ) -> int:
         cnx = get_sql_connection()
