+++ api/persistence/stubs/washroom_stub.py
@@ -19,7 +19,7 @@ class WashroomsStubPersistence(IWashroomsPersistence):
         floor: int,
         gender: str,
         amenities_id: int,  # Foreign Key
-        overall_rating: int,
+        overall_rating: float,
         average_ratings_id: int  # Foreign Key
     ) -> int:
         washroom_id = len(self.washrooms)
